Rename File CountlyMetrics -> MetricsProvider
