+++ src/index.ts
@@ -1,2 +1,2 @@
 export * from './components/index.js'
-export * from './CountlyMetrics.js'
+export * from './MetricsProvider.js'
